move recreate command to archiver.recreate
