+++ setup.cfg
@@ -126,6 +126,7 @@ per_file_ignores =
     src/borg/archiver/help.py:E501,F405
     src/borg/archiver/keys.py:F405
     src/borg/archiver/prune.py:F405
+    src/borg/archiver/recreate.py:F405
     src/borg/archiver/rdelete.py:F405
     src/borg/archiver/rlist.py:E501
     src/borg/archiver/tar.py:F405
